fixed small bug with saving templates

+++ src/commandsHandlers.ts
@@ -4,7 +4,7 @@ import * as fs from 'fs'
 import * as afs from 'node:fs/promises'
 import path from "path"
 import color from 'picocolors'
-import { copyDirectoryContents, getFilesIgnoredByGit } from "./utils.js"
+import { CopyDirectoryContentsParams, copyDirectoryContents, getFilesIgnoredByGit } from "./utils.js"
 import { S_BAR } from './clack/styled/utils.js'
 import { WORK_DIR } from './index.js'
 import { ScriptsFunction, TEMPLATES_PATH } from './constants.js'
@@ -13,11 +13,16 @@ interface SaveOptions {
     templateName: string
     sourceRelativePath: string
     withGitignore: boolean
+    templateDescsription?: string
 }
 
-export async function save({ templateName, sourceRelativePath, withGitignore }: SaveOptions) {
+export async function save({ templateName, sourceRelativePath, withGitignore, templateDescsription }: SaveOptions) {
     const templatePath = `${TEMPLATES_PATH()}/${templateName}`
     const sourceAbsolutePath = `${WORK_DIR}/${sourceRelativePath}`
+    const copyArgs: CopyDirectoryContentsParams = {
+        fromPath: sourceAbsolutePath,
+        toPath: templatePath,
+    }
 
     if (!fs.existsSync(sourceAbsolutePath)) {
         p.log.error(`Directory specified as source: ${sourceAbsolutePath} does not exist`)
@@ -30,35 +35,31 @@ export async function save({ templateName, sourceRelativePath, withGitignore }: 
             message: `Template ${templateName} already exists. Override It?`,
             initialValue: false
         }) as boolean
-        fs.rmSync(templatePath, { recursive: true, force: true })
+
+        if (override) fs.rmSync(templatePath, { recursive: true, force: true })
+    }
+
+    let ignoredFiles: string[] = []
+    if (withGitignore) {
+        try {
+            ignoredFiles = await getFilesIgnoredByGit()
+            copyArgs.validate = ({ createName, createPath, isFile, sourcePath }) => {
+                if (ignoredFiles.includes(createName)) {
+                    p.log.info(`${color.dim(createName)} ignored by git. Skipping It...`)
+                    return false
+                }
+                return true
+            }
+        } catch (e) {
+            p.log.warn(`Problem with checking git files: ${color.dim((e as Error).message)}`)
+        }
     }
 
     if (!fs.existsSync(templatePath) || override) {
         await afs.mkdir(templatePath, { recursive: true })
 
-        if (withGitignore) {
-            let files: string[] = []
-            try {
-                files = await getFilesIgnoredByGit()
-            } catch (e) {
-                p.log.warn(`Problem with checking git files: ${color.dim((e as Error).message)}`)
-            }
-            copyDirectoryContents(sourceAbsolutePath, templatePath, {
-                validate: ({ createName, createPath, isFile, sourcePath }) => {
-                    if (files.includes(createName)) {
-                        p.log.info(`${color.dim(createName)} ignored by git. Skipping It...`)
-                        return false
-                    }
-                    return true
-                }
-            })
-
-            p.log.success(`Saved ${color.cyan(templateName)} template at ${color.dim(templatePath)}`)
-        }
-        else {
-            copyDirectoryContents(sourceAbsolutePath, templatePath)
-            p.log.success(`Saved ${color.cyan(templateName)} template at ${color.dim(templatePath)}`)
-        }
+        copyDirectoryContents(copyArgs)
+        p.log.success(`Saved ${color.cyan(templateName)} template at ${color.dim(templatePath)}`)
     }
     else {
         p.log.error(`Aborted saving ${color.cyan(templateName)} template.`)
@@ -85,7 +86,7 @@ export function paste({ templateName, destinationRelativePath }: PasteOptions) {
         p.log.info(`Directory ${color.dim(destinationAbsolutePath)} does not exists. Creating it...`)
     }
 
-    copyDirectoryContents(templatePath, destinationAbsolutePath)
+    copyDirectoryContents({ fromPath: templatePath, toPath: destinationAbsolutePath })
     p.log.success(`Pasted ${templateName} template to ${color.dim(destinationAbsolutePath)}`)
 }
